Attempt to address concern about lack of evidence linked.
It's not our goal here to document all the irregularities in this
election.  The original draft included a last-minute request from a
frustrated party to communicate that.  With the LWN link, I think we
have what we need in this regard anyway.

NOTE (added on 2025-04-17):

This ‘drafting’ branch contains detailed commits during the drafting
process.  During that drafting, research about what occurred during
the 2025 OSI election was ongoing.  As such, some statements in this
commit may be inaccurate.  Please see the ‘main’ branch for the
official, released version of the petition.