add sentence about misrepresenting the electorate
and some other tweaks

NOTE (added on 2025-04-17):

This ‘drafting’ branch contains detailed commits during the drafting
process.  During that drafting, research about what occurred during
the 2025 OSI election was ongoing.  As such, some statements in this
commit may be inaccurate.  Please see the ‘main’ branch for the
official, released version of the petition.