Add my signature
I believe this untransparent behaviour, whether the result of a mistake or otherwise, is harmful for people's trust in and the integrity of the OSI and should be reversed as soon as possible.

My understanding is that the bylaws make it clear that the OSI is not a pure democracy and that the elections aren't binding, but also don't suggest such a misleading situation where people are on the ballots but don't have their votes made public, and where unbeforementioned deadlines pop up. Transparency of all votes is my demand, even if not signing the form remains an exclusion criterion.

Such transparency would give the OSI's rejection of "trolls" who claim the actions were to avoid Free Software ideals reaching the OSI more justifiability.

My primary demand is transparency, and my wish is more democracy.