Signature with extra statement in git commit message
I've had concerns for a while about the direction the #OSI has been
going with AI for a while.  The publishing of the #OSAID weakens the

I'm not a member but the definition does require community participation
in terms of the licences we choose and how we advocate for Open Source.

I've have advocated quietly for Open Source for decades now.  I consider
myself friends with some of the members.

This is not a petition I sign lightly, it may after all have an affect
on my career and my standing in the community.

But on occasion sacrifices must be made, and I've watched with concern
the direction the OSI seems to be going in.

The recent board elections showed a disturbing pattern of
obfuscation.  We need to be able to trust Organisations that are meant
to safeguard the definition of what Open source is.

It's surprising that the #OSI decided to obfuscate the election results.

The election is done, but the OSI has damaged our trust.

A way to restore that trust for the community would be to publish the
voting tallies in full.

The reform candidates wanted to repair the relationship the OSI has with
the greater community.  If you're a member of the OSI or of the Greater
FOSS community this petition is the place to sign.

Because if nothing else, it's a way to signal you don't consent to the
direction the OSI is going.  You don't consent to the damage that OSAID
is doing to the Open Source Definition.

While some on the board may see this as a number of troublemakers and
enemies.  I'd urge them to see this as another way.

It's a call for a more open organisation that is willing to listen to
not just it's members but the wider community that it's meant to guide.
It's a call to talk with the community again and reset.

Esther Payne
Onepict.com