Add signature.
I'm for free software and most probably against Open Source, however
as I understand OSI's goal is to promote Open Source, which they then
also define, so because of that I've also a hard time understanding
the OSI's goals, so it's also hard to know for sure.

I'm for users freedom for both individual and organizations, and so in
general my positions align more with the free software movement than
the open source movement.

That said, both community are intertwined together, and I also
contributed to many Open Source projects as well (I left politics
aside to do that), in order to advance users freedom.

In addition, as a user and contributor, I'm also affected by OSI
policies in an indirect way, and I've always tried to find ways to
make people involved in users freedom or other practical goals that
are not in direct conflict with it collaborate together (though I
strongly prefer if this is done on our terms).

So I also find it important that people that don't care about end users
freedom also have a voice and I don't want to interfere with that at
all.

So because of that I want to sign a bit like an Amicus curiae because
while I'm probably not aligned with Open Source's goals, I am affected
by decisions made by the OSI through my contributions and because of
that I need it to have clear process and goals, even if they are in
opposition to my own user freedom goals, to be able to understand
where or how I can collaborate or not with the Open Source movement.

Since the goal of this petition is to increase clarity, I want to sign
it because of that.

Also note that I've also done a bit of research before signing this
petition in order to understand if there was a risk of interference
with the OSI mission.

And my understanding is that increasing clarity or even having Bradley
Kuhn and Richard Fontana elected would in fact make things more clear
and not go against the OSI's mission (assuming that OSI does not care
about users freedom and instead promotes other direct or indirect
benefits of publishing under free licenses), so I can safely sign
without putting the OSI at risk of interference.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>