Sign the petition (Faidon Liambotis)
For the record, the Board Agreement was instituted during my term, and I
voted in favor of it, a decision I stand by.

I continue to believe that candidates who decline to meet essential
requirements of the job should not be seated on the board. At the same
time, I believe that the manner in which the OSI chose to enforce and
communicate this requirement was inappropriate and has eroded trust in
the organization.

In other words, this signature should not be misconstrued as a statement
against the Board Agreement or any other position the affected
candidates may have held (such as the OSAID), but solely as a statement
regarding the election process itself.

Signed-off-by: Faidon Liambotis <paravoid@debian.org>