[refactor] Signature: Brian C. Arnold - personal statement:
It is imperative that any organization that participates[1] in
"community-building, education, and public advocacy to promote [..]
non-proprietary software" must recognize that having an arbitrarily
decided and unpredictable election process is antithetical the very
principles of openness and freedom for which it claims to stand.

Under the Open Source Definition[2], the OSI is quite clear that
licenses must be Technology-Neutral, and as such, OSS software can
be written for any technology stack and utilize the same OSI
license, and what is an election except software running on the
oldest technology stack known to humanity: Civilization.

To that end, the OSI must recognize that the recent election process
was not published in source form, but effectively only in compiled
form. Requirements and demands of the participants were changed
during the election, without sufficient explanation of the supposed
original process, or a history of the changes.

Please release your election algorithm, not just STV, but the entire
process, and under a true OSI approved license.

1.https://opensource.org/about
2.https://opensource.org/osd